Move test to correct file and change the comparisor to the correct one
